Refactor file name and move helper code to the right place
